requirements.txt
python-telegram-bot>=20.7
anthropic>=0.20.0
APScheduler>=3.10.4
aiohttp>=3.9.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.0
aiosqlite>=0.19.0
lxml>=5.1.0
